+++ src/AppVersionFacade.php
@@ -1,18 +1,19 @@
 <?php
 
+declare(strict_types = 1);
+
 namespace AvtoDev\AppVersion;
 
-use Illuminate\Support\Facades\Facade;
 use AvtoDev\AppVersion\Contracts\AppVersionManagerContract;
 
-class AppVersionFacade extends Facade
+class AppVersionFacade extends \Illuminate\Support\Facades\Facade
 {
     /**
      * Get the registered name of the component.
      *
      * @return string
      */
-    protected static function getFacadeAccessor()
+    protected static function getFacadeAccessor(): string
     {
         return AppVersionManagerContract::class;
     }