+++ tests/Bootstrap/bootstrap.php
@@ -1,3 +1,0 @@
-<?php
-
-new \AvtoDev\AppVersion\Tests\Bootstrap\TestsBootstrapper;